chore: Functional split components
