+++ src/app/(uc)/signin/components/LoginTypeSwitcher.js
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2024 OpenBuild
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use client'
+
+import { SvgIcon } from '@/components/Image'
+
+export default function LoginTypeSwitcher({ loginType, handleChangeLoginType }) {
+	return (
+		<p className="flex mt-4 text-sm">
+			Switch to {loginType === 'verifyCode' ? 'password' : 'verify code'} login
+			<SvgIcon
+				name="change"
+				size={14}
+				className="ml-2 cursor-pointer"
+				onClick={handleChangeLoginType}
+			/>
+		</p>
+	)
+}
